third_party/blink/web_tests/platform/generic/fast/dom/navigation-type-navigate-expected.txt

This tests that performance.navigation.type is 0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigation.type is 0
PASS successfullyParsed is true

TEST COMPLETE
